feat(core): add omnitracking's checkout started event mappings
- Added checkout started omnitracking event mapping.
